added additional themeing to site
i had to insert a disable eslint comment line 37 src/components/layout.js, it is stating that theme is not used when it is, not sure why it is doing that. My next commit should be a basic codesandbox component, I have been playing with ways to do it but Ill commit the additional styling for now.