Added general purpose Dockerfile
The point of this file is to guarentee a workable environment for the project. It can be useful to build this docker image and then mount the project as a volumne. The Dockerfile has some specific instructions about going about this. There is also a .dockerignore, its not particularly useful yet, but it may have use later down the road.